+++ packages/example/electron.vite.config.ts
@@ -7,12 +7,21 @@ export default defineConfig({
   },
   preload: {
     plugins: [externalizeDepsPlugin()],
+    build: {
+      rollupOptions: {
+        input: {
+          index: resolve("src/preload/index.ts"),
+          receiver: resolve("src/preload/receiver.ts"),
+        },
+      },
+    },
   },
   renderer: {
     build: {
       rollupOptions: {
         input: {
           index: resolve("src/renderer/index.html"),
+          "receiver-test": resolve("src/renderer/receiver-test.html"),
         },
       },
     },